+++ Cargo.toml
@@ -34,6 +34,7 @@ members = [
     "src/mito",
     "src/object-store",
     "src/partition",
+    "src/prometheus",
     "src/promql",
     "src/query",
     "src/script",
