+++ src/frontend/src/instance/prometheus.rs
@@ -22,9 +22,9 @@ use common_query::Output;
 use common_recordbatch::RecordBatches;
 use common_telemetry::logging;
 use metrics::counter;
+use prometheus::{self, Metrics};
 use prost::Message;
 use servers::error::{self, Result as ServerResult};
-use servers::prometheus::{self, Metrics};
 use servers::query_handler::grpc::GrpcQueryHandler;
 use servers::query_handler::{PrometheusProtocolHandler, PrometheusResponse};
 use session::context::QueryContextRef;
