+++ src/prometheus/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod error;
+mod prometheus;
+
+pub use crate::prometheus::*;
