+++ src/servers/Cargo.toml
@@ -46,12 +46,12 @@ metrics-process = "<1.0.10"
 mime_guess = "2.0"
 num_cpus = "1.13"
 once_cell = "1.16"
-openmetrics-parser = "0.4"
 opensrv-mysql = "0.4"
 parking_lot = "0.12"
 pgwire = "0.14.1"
 pin-project = "1.0"
 postgres-types = { version = "0.2", features = ["with-chrono-0_4"] }
+prometheus = { path = "../prometheus" }
 promql-parser = "0.1.1"
 prost.workspace = true
 query = { path = "../query" }
