+++ src/servers/src/lib.rs
@@ -32,7 +32,6 @@ pub mod mysql;
 pub mod opentsdb;
 pub mod postgres;
 pub mod prom;
-pub mod prometheus;
 pub mod query_handler;
 pub mod server;
 mod shutdown;
