+++ src/servers/tests/http/prometheus_test.rs
@@ -24,12 +24,11 @@ use axum_test_helper::TestClient;
 use common_query::Output;
 use common_test_util::ports;
 use datatypes::schema::Schema;
+use prometheus::{snappy_compress, Metrics};
 use prost::Message;
 use query::parser::PromQuery;
 use servers::error::{Error, Result};
 use servers::http::{HttpOptions, HttpServerBuilder};
-use servers::prometheus;
-use servers::prometheus::{snappy_compress, Metrics};
 use servers::query_handler::grpc::GrpcQueryHandler;
 use servers::query_handler::sql::SqlQueryHandler;
 use servers::query_handler::{PrometheusProtocolHandler, PrometheusResponse};
