+++ tests-integration/Cargo.toml
@@ -34,6 +34,7 @@ meta-srv = { path = "../src/meta-srv" }
 mito = { path = "../src/mito", features = ["test"] }
 object-store = { path = "../src/object-store" }
 once_cell = "1.16"
+prometheus = { path = "../src/prometheus" }
 query = { path = "../src/query" }
 rand.workspace = true
 rstest = "0.17"
